updating my fork (#1)
* Adding Betrayal at House on the Hill rulebook.

* create colossal-arena rulebook

hello, saw your post on reddit and wanted to contribute this classic Knizia game! I pretty much ripped the original rulebook text minus the images. Hope that's ok. Thanks!

* Moved colossal arena

* moved to md

* Minor tweaks

* added colossal arena to recents

* Add a rulebook for Jaipur

* Create between-two-cities.md

I am in the process of writing up summarised versions of many games - the idea being for use as refreshers or teaching aids. I understand this is not exactly what your site is doing, but take a look and maybe there is a place for it - differently tagged, or just as it is. If not, feel free to ignore.

Contents and tags are stubbed for now.

This game is Between Two Cities: https://boardgamegeek.com/boardgame/168435/between-two-cities

* Add tags. Add content. Minor edits.

* Create snow-tails.md

* Update between-two-cities.md

* Add link to site. Remove 'refresher' tag.

* Update the-resistance.md

Fixed player chart

* Remove "Notes" section

* Create mamma-mia.md

* Update how-to-help.md

* Create bang-the-dice-game.md

* Update mamma-mia.md

* Update bang-the-dice-game.md

* Create mansions-of-madness-second-edition.md

* Update mansions-of-madness-second-edition.md